graphics.hidd: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@41318 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/hidds/graphics/sync.c
@@ -162,15 +162,15 @@ OOP_Object *Sync__Root__New(OOP_Class *cl, OOP_Object *o, struct pRoot_New *msg)
 		    	switch (c)
 		        {
 		    	case 'b':
-		            l = snprintf(d, dlen, "%u", board);
+		            l = snprintf(d, dlen, "%u", (unsigned)board);
 		            break;
 
 		    	case 'h':
-		            l = snprintf(d, dlen, "%u", data->hdisp);
+		            l = snprintf(d, dlen, "%u", (unsigned)data->hdisp);
 			    break;
 
 		    	case 'v':
-		            l = snprintf(d, dlen, "%u", data->vdisp);
+		            l = snprintf(d, dlen, "%u", (unsigned)data->vdisp);
 			    break;
 
 		    	default:
